Rename use_cases to use_case
