+++ src/main/java/use_case/DataChangeListener.java
@@ -1,4 +1,4 @@
-package use_cases;
+package use_case;
 
 import entity.VaultItem;
